refactor(modern_bpf): use just one file for enter and exit event
